@@ -85,7 +85,7 @@ namespace PortfolioPicker.App
                 "## stats",
                 Row("stat", "value"),
                 Row("---", "---"),
-                Row("total value of assets", string.Format("{0:c}", TotalValue)),
+                Row("total value of assets", string.Format("${0:0.00}", TotalValue)),
                 Row("expense ratio", string.Format("{0:0.0000}", ExpenseRatio)),
                 Row("percent stocks", string.Format("{0:0.0}%", ExposureRatios.Percent(AssetClass.Stock))),
                 Row("percent bonds", string.Format("{0:0.0}%", ExposureRatios.Percent(AssetClass.Bond))),
@@ -116,7 +116,7 @@ namespace PortfolioPicker.App
                     foreach (var p in a.Positions.OrderByDescending(x => x.Value))
                     {
                         var f = Fund.Get(p.Symbol);
-                        lines.Add(Row(a.Name, Url(p.Symbol), string.Format("{0:c}", p.Value, f.Description)));
+                        lines.Add(Row(a.Name, Url(p.Symbol), string.Format("${0:0.00}", p.Value), f.Description));
                     }
                 }
             }
